==> 5_sm_data/1_raw_ERA5_Land.ipynb <==
### Processing ERA5-Land data

Load CDO module -> use that to change from grib to netcdf and to resample to 0.5 degree resolution.  
Using remapcon since it's a conservative remapping technique (area-averaged mean).

import os

from cdo import Cdo
cdo = Cdo()  # Initialize CDO object

downloads_folder = os.path.join(os.path.expanduser("~"), "Downloads")
# Define input and output file paths
input_file = os.path.join(downloads_folder, "797bb28b9ae1758bdf172d3035541ef.grib")
output_file = os.path.join(downloads_folder, "output_remapped.grib")
grid_file = os.path.join(downloads_folder, "grid.txt")

# Check if NetCDF file already exists
if not os.path.exists(output_file):
    print("Remapped file not found. Creating it now...")
    # Run the remapping
    cdo.remapcon(grid = grid_file, input=input_file, output=output_file)
    print(f"Created: {output_file}")
else:
    print(f"Remapped file already exists: {output_file}")

Converting the grib to netcdf

# Define input and output file paths
input_file = os.path.join(downloads_folder, "output_remapped.grib")
output_file = os.path.join(downloads_folder, "output_remapped.nc")

# Check if NetCDF file already exists
if not os.path.exists(output_file):
    print("NetCDF file not found. Creating it now...")
    #Convert grib output to netcdf
    cdo.copy(input=input_file, output=output_file, options='-f nc')
    print(f"Created: {output_file}")
else:
    print(f"NetCDF file already exists: {output_file}")
